+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 *Revisions*
 
 * Updated the Witch's description and rules. ([#138](https://github.com/hjwylde/werewolf/issues/138))
+* Updated the Scapegoat's description. ([#133](https://github.com/hjwylde/werewolf/issues/133))
 
 ### v0.4.11.0